moved project card into seperate component
