+++ src/routes/(authenticated)/projects/+page.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
 	import { MainContent, MainContentHeader } from '$components/layout/main-content'
-	import ProjectCards from '$components/container/projects/project-cards.svelte'
 	import CreateProject from '$components/container/projects/create-project.svelte'
 	import type { PageData } from './$types'
+	import ProjectCardsLayout from '$components/container/projects/project-cards-layout.svelte'
 
 	export let data: PageData
 </script>
@@ -11,5 +11,5 @@
 	<MainContentHeader title="Projects">
 		<CreateProject data={data.form} />
 	</MainContentHeader>
-	<ProjectCards projects={data.projects} />
+	<ProjectCardsLayout projects={data.projects} />
 </MainContent>
